fix: port to create-react-app
Taquito, integration library for Tezos blockchain, seems to be using CommonJS require statements in
its code. That is incompatible with Vite and could not be fixed with a rollup plugin. So, it became
necessary to migrate to a webpack-based create-react-app setup, that too with a react-app-rewired
custom webpack config to polyfill Node core modules.